Preliminary code to add links at OPI runtime.
Previously you could only update links from the share_link option
at CSS startup.  In order to use the launcher satisfactorily, we
need to be able to update links when running an OPI on an existing
CSS session.

This commit allows you to do this, but it may not be the cleanest way of
doing so.